+++ subpackages/cli/source/odood/cli/commands/database.d
@@ -5,6 +5,8 @@ private import std.stdio;
 private import std.format: format;
 private import std.exception: enforce;
 private import std.typecons;
+private import std.algorithm;
+private import std.string;
 
 private import thepath: Path;
 private import commandr: Argument, Option, Flag, ProgramArgs;
@@ -308,6 +310,48 @@ class CommandDatabaseStun: OdoodCommand {
 }
 
 
+class CommandDatabaseListInstalledAddons: OdoodCommand {
+    this() {
+        super(
+            "list-installed-addons",
+            "List addons installed in specified databases");
+        this.add(new Option(
+            "d", "db", "Name of database to to check for addons.").repeating);
+        this.add(new Option(
+            "o", "out-file", "Path to file where to store generated requirements"));
+        this.add(new Flag(
+            "a", "all", "Check all databases"));
+    }
+
+    public override void execute(ProgramArgs args) {
+        auto project = Project.loadProject;
+
+        Path output_path;
+        bool print_to_file = false;
+        if (args.option("out-file")) {
+            output_path = Path(args.option("out-file"));
+            print_to_file = true;
+        }
+
+        string[] dbnames = args.flag("all") ? project.databases.list() : args.options("db");
+
+        string[] addon_names;
+        foreach(dbname; dbnames) {
+            auto db = project.dbSQL(dbname);
+            scope(exit) db.close;
+
+            auto res = db.runSQLQuery("SELECT array_agg(name) FROM ir_module_module WHERE state = 'installed'")[0][0].as!(string[]).get;
+            addon_names ~= res;
+        }
+        string result = addon_names.sort.uniq.join("\n") ~ "\n";
+        if (print_to_file)
+            output_path.writeFile(result);
+        else
+            writeln(result);
+
+    }
+}
+
 class CommandDatabase: OdoodCommand {
     this() {
         super("db", "Database management commands");
@@ -320,6 +364,7 @@ class CommandDatabase: OdoodCommand {
         this.add(new CommandDatabaseBackup());
         this.add(new CommandDatabaseRestore());
         this.add(new CommandDatabaseStun());
+        this.add(new CommandDatabaseListInstalledAddons());
     }
 }